move utils up
