+++ src/api/invoices/msw-handlers.tsx
@@ -1,6 +1,6 @@
 import { delay, http, HttpResponse } from 'msw';
 
-import { Invoice } from '@/app/invoices/components/schema';
+import { Invoice } from '@/app/invoices/schema';
 import { env } from '@/lib/env';
 
 import { sampleInvoices } from './sample-invoices';
