+++ src/app/invoices/components/dashboard/invoice-sheet.tsx
@@ -21,11 +21,10 @@ import {
 import { formatCurrency } from '@/lib/currency';
 import { cn } from '@/lib/utils';
 
-import { Invoice } from '../schema';
+import { Invoice } from '../../schema';
+import { checkOverdue, formatDate, InvoiceStatus } from '../../utils';
 import { InvoiceSheetCancelDrawer } from './invoice-sheet-cancel-drawer';
 import { InvoiceStatusBadge } from './invoice-status-badge';
-import { checkOverdue, InvoiceStatus } from './utils';
-import { formatDate } from './utils';
 
 // TODO: Is there a better scroll solution? Perhaps with sticky header and footer?
 // TODO: Use custom easing curves to improve the feel of the sheet open animation (match vaul?)
