+++ src/app/invoices/components/deprecated/create/invoice-details.tsx
@@ -26,8 +26,8 @@ import { Input } from '@/components/ui/input';
 import { formatCurrency } from '@/lib/currency';
 import { cn } from '@/lib/utils';
 
-import { calculateSubtotalTaxAndTotal } from '../../dashboard/utils';
-import { invoiceFormStringValidator } from '../../schema';
+import { invoiceFormStringValidator } from '../../../schema';
+import { calculateSubtotalTaxAndTotal } from '../../../utils';
 import { CreateInvoiceFooter } from './create-invoice-footer';
 import { CreateInvoiceHeader } from './create-invoice-header';
 import { CreateInvoiceShell } from './create-invoice-shell';
