+++ src/app/invoices/create/page.tsx
@@ -6,7 +6,7 @@ import { createInvoice } from '@/api/invoices';
 import { Authenticated } from '@/app/authenticated';
 
 import { CreateInvoice } from '../components/create-invoice';
-import { InvoiceForm } from '../components/schema';
+import { InvoiceForm } from '../schema';
 
 export default function CreateInvoicePage() {
   const router = useRouter();
