+++ src/app/invoices/hooks/use-invoice-form.ts
@@ -1,6 +1,6 @@
 import { useFormContext } from 'react-hook-form';
 
-import { InvoiceForm } from '../components/schema';
+import { InvoiceForm } from '../schema';
 
 /** Hook for accessing the invoice form context */
 export const useInvoiceForm = () => useFormContext<InvoiceForm>();
