+++ src/app/invoices/hooks/use-invoice-printer.tsx
@@ -3,7 +3,7 @@ import { useRef } from 'react';
 import { useReactToPrint } from 'react-to-print';
 
 import { InvoiceDocument } from '../components/invoice-document';
-import { Invoice } from '../components/schema';
+import { Invoice } from '../schema';
 
 /**
  * Use this hook to print an invoice in a place where it's not visible to the user.
