+++ src/app/invoices/schema.ts
@@ -2,7 +2,8 @@ import { addDays, startOfDay } from 'date-fns';
 import { z } from 'zod';
 
 import { checkInvoiceNumber } from '@/api/invoices';
-import { InvoiceStatus } from '@/app/invoices/components/dashboard/utils';
+
+import { InvoiceStatus } from './utils';
 
 // TODO: Note there are duplicates of v1 invoice stuff here. At the end of the day, there should only be one
